requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-jazzmin==3.0.1
django-js-asset==2.2.0
idna==3.10
pillow==11.0.0
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
